switched stake batch lock from simple bool to enum in preparation of updating the staking workflow
to make the staking workflow more robust to handle failures and not cause the state to be inconsistent with the staking pool. For example, if callback gas is configured too low for on_deposit_and_stake, then it will fail even though the batch funds have been transferred over to the staking pool. The problem is that the stake batch remains in place and the contract loses the information that the stake batch funds have already been staked